+++ db/src/orm/java/net/gdface/facelog/dborm/image/FlStoreManager.java
@@ -159,12 +159,16 @@ public class FlStoreManager implements TableManager<FlStoreBeanBase,FlStoreBean>
      * Loads a {@link FlStoreBean} from the fl_store using primary key fields.
      *
      * @param md5 String - PK# 1
-     * @return a unique FlStoreBean or {@code null} if not found
+     * @return a unique FlStoreBean or {@code null} if not found or have null argument
      * @throws DAOException
      */
     //1
+    @SuppressWarnings("unused")
     public FlStoreBean loadByPrimaryKey(String md5) throws DAOException
     {
+        if(null == md5){
+            return null;
+        }
         Connection c = null;
         PreparedStatement ps = null;
         try
@@ -199,7 +203,7 @@ public class FlStoreManager implements TableManager<FlStoreBeanBase,FlStoreBean>
      * when you don't know which is primary key of table,you can use the method.
      * @author guyadong
      * @param bean the {@link FlStoreBean} with primary key fields
-     * @return a unique {@link FlStoreBean} or {@code null} if not found
+     * @return a unique {@link FlStoreBean} or {@code null} if not found or bean is null
      * @throws DAOException
      * @see {@link #loadByPrimaryKey(String md5)}
      */
@@ -238,15 +242,20 @@ public class FlStoreManager implements TableManager<FlStoreBeanBase,FlStoreBean>
     }
     
     /**
-     * Delete row according to its primary keys.
-     *
+     * Delete row according to its primary keys.<br>
+     * all keys must not be null
+     * 
      * @param md5 String - PK# 1
      * @return the number of deleted rows
      * @throws DAOException
      */
     //2
+    @SuppressWarnings("unused")
     public int deleteByPrimaryKey(String md5) throws DAOException
     {
+        if(null == md5){
+            throw new IllegalArgumentException("primary keys must no be null ");
+        }
         Connection c = null;
         PreparedStatement ps = null;
         try
@@ -1566,7 +1575,7 @@ public class FlStoreManager implements TableManager<FlStoreBeanBase,FlStoreBean>
     public int deleteUsingTemplate(FlStoreBeanBase beanBase) throws DAOException
     {
         FlStoreBean bean=FlStoreBeanBase.toFullBean(beanBase);
-        if (bean.isMd5Initialized()) {
+        if(bean.isMd5Initialized() && null != bean.getMd5()){
             return this.deleteByPrimaryKey(bean.getMd5());
         }
         Connection c = null;
