+++ db/src/orm/java/net/gdface/facelog/dborm/log/FlLogManager.java
@@ -208,12 +208,16 @@ public class FlLogManager implements TableManager<FlLogBeanBase,FlLogBean>
      * Loads a {@link FlLogBean} from the fl_log using primary key fields.
      *
      * @param id Integer - PK# 1
-     * @return a unique FlLogBean or {@code null} if not found
+     * @return a unique FlLogBean or {@code null} if not found or have null argument
      * @throws DAOException
      */
     //1
+    @SuppressWarnings("unused")
     public FlLogBean loadByPrimaryKey(Integer id) throws DAOException
     {
+        if(null == id){
+            return null;
+        }
         Connection c = null;
         PreparedStatement ps = null;
         try
@@ -248,7 +252,7 @@ public class FlLogManager implements TableManager<FlLogBeanBase,FlLogBean>
      * when you don't know which is primary key of table,you can use the method.
      * @author guyadong
      * @param bean the {@link FlLogBean} with primary key fields
-     * @return a unique {@link FlLogBean} or {@code null} if not found
+     * @return a unique {@link FlLogBean} or {@code null} if not found or bean is null
      * @throws DAOException
      * @see {@link #loadByPrimaryKey(Integer id)}
      */
@@ -287,15 +291,20 @@ public class FlLogManager implements TableManager<FlLogBeanBase,FlLogBean>
     }
     
     /**
-     * Delete row according to its primary keys.
-     *
+     * Delete row according to its primary keys.<br>
+     * all keys must not be null
+     * 
      * @param id Integer - PK# 1
      * @return the number of deleted rows
      * @throws DAOException
      */
     //2
+    @SuppressWarnings("unused")
     public int deleteByPrimaryKey(Integer id) throws DAOException
     {
+        if(null == id){
+            throw new IllegalArgumentException("primary keys must no be null ");
+        }
         Connection c = null;
         PreparedStatement ps = null;
         try
@@ -1098,23 +1107,6 @@ public class FlLogManager implements TableManager<FlLogBeanBase,FlLogBean>
 
             ps.executeUpdate();
 
-            if (!bean.isIdModified())
-            {
-                PreparedStatement ps2 = null;
-                ResultSet rs = null;
-                try {
-                    ps2 = c.prepareStatement("SELECT last_insert_id()");
-                    rs = ps2.executeQuery();
-                    if(rs.next()) {
-                        bean.setId(Manager.getInteger(rs, 1));
-                    } else {
-                        this.getManager().log("ATTENTION: Could not retrieve generated key!");
-                    }
-                } finally {
-                    this.getManager().close(ps2, rs);
-                }
-            }
-
             bean.isNew(false);
             bean.resetIsModified();
             this.afterInsert(bean); // listener callback
@@ -1633,7 +1625,7 @@ public class FlLogManager implements TableManager<FlLogBeanBase,FlLogBean>
     public int deleteUsingTemplate(FlLogBeanBase beanBase) throws DAOException
     {
         FlLogBean bean=FlLogBeanBase.toFullBean(beanBase);
-        if (bean.isIdInitialized()) {
+        if(bean.isIdInitialized() && null != bean.getId()){
             return this.deleteByPrimaryKey(bean.getId());
         }
         Connection c = null;
